test: extract migration tests from state.test.ts into state-migration.test.ts
Move StateManager v1→v2 migration, legacy state cleanup, issueListPath
config, migrateFromLegacyLocation, migrateV1ToV2 fallback branches, and
schema validation describe blocks into a dedicated file. Remove now-unused
acquireLock/releaseLock/atomicWriteFileSync imports from state.test.ts.

Total: 200 tests across 4 files (unchanged).
